requirements.txt
mcp[cli]
mwclient
python-dotenv
uvicorn
pytest
